Move static extension
