+++ src/static-extension/StaticExtension.js
@@ -1,6 +1,6 @@
-import {ASSETS_KEY} from './runtime-contants';
-import {getModule} from './utils/get-module';
-import {getRegexMatches} from './utils/get-regex-matches';
+import {ASSETS_KEY} from './contants';
+import {getModule} from '../utils/get-module';
+import {getRegexMatches} from '../utils/get-regex-matches';
 
 export class StaticExtension {
     constructor() {
