+++ src/static-extension/get-static-extension.js
